Move summarizeTimeEntries into harvest.ts.
