+++ src/lib/harvest.ts
@@ -1,7 +1,7 @@
 import 'whatwg-fetch';
 
 import { getAuthToken, getEnableConfig } from './store';
-import { Project, TaskAssignment, TimeEntry } from './types';
+import { Project, TaskAssignment, TaskSummaries, TimeEntry } from './types';
 
 export const API_BASE_URL = 'https://api.harvestapp.com/v2/';
 
@@ -72,3 +72,15 @@ export const getTimeEntries = (
 
 export const getProjects = (t: Trello): Promise<Project[]> =>
   getHarvestData(t, 'projects?is_active=true', 'projects');
+
+// summarize an array of TimeEntry to a task-id -> TimeSummary map
+// where a TimeSummary is a dev-name -> total-hours map
+export const summarizeTimeEntries = (entries: TimeEntry[]): TaskSummaries =>
+  entries.reduce(
+    (acc, entry) => {
+      const summary = (acc[entry.task.id] = acc[entry.task.id] || {});
+      summary[entry.user.name] = (summary[entry.user.name] || 0) + entry.hours;
+      return acc;
+    },
+    {} as { [key: number]: { [key: string]: number } },
+  );
