+++ test/harvest.test.js
@@ -4,6 +4,7 @@ import {
   API_BASE_URL,
   getHarvestData,
   getHarvestJSON,
+  summarizeTimeEntries,
 } from '../src/lib/harvest';
 import trelloMock from './trello-mock.js';
 
@@ -34,5 +35,18 @@ describe('getHarvestData', () => {
       .getOnce(secondUrl, { things: ['three'], links: { next: null } });
     const data = await getHarvestData(trelloMock, 'some/path', 'things');
     expect(data).toEqual(['one', 'two', 'three']);
+  });
+});
+
+describe('summarizeTimeEntries', () => {
+  test('summarizes all given time entries', () => {
+    const entries = [
+      { task: { id: 1 }, user: { name: 'A' }, hours: 1.2 },
+      { task: { id: 2 }, user: { name: 'A' }, hours: 2.4 },
+      { task: { id: 2 }, user: { name: 'A' }, hours: 0.2 },
+      { task: { id: 2 }, user: { name: 'B' }, hours: 0.7 },
+    ];
+    const summaries = summarizeTimeEntries(entries);
+    expect(summaries).toEqual({ 1: { A: 1.2 }, 2: { A: 2.6, B: 0.7 } });
   });
 });
